move SMILES in the AST
