--- documentation/source/api/smiles.rst
@@ -1,5 +1,0 @@
-SMILES (``osmipy.smiles``)
-==========================
-
-.. automodule:: osmipy.smiles
-    :members:
