+++ documentation/source/api/smiles_visitors.rst
@@ -1,0 +1,5 @@
+SMILES visitors (``osmipy.smiles_visitors``)
+============================================
+
+.. automodule:: osmipy.smiles_visitors
+    :members:
