+++ tests/tests_lexer.py
@@ -1,6 +1,6 @@
 from tests import OSmiPyTestCase
 
-from osmipy import smiles, lexer
+from osmipy import lexer, tokens
 from osmipy.tokens import Token as Tok
 
 
@@ -10,33 +10,33 @@ class LexerTestCase(OSmiPyTestCase):
         """Test that the lexer gives the good tokens"""
 
         tests = [
-            ('cccccc', [Tok(smiles.ATOM, 'c')] * 6),
+            ('cccccc', [Tok(tokens.ATOM, 'c')] * 6),
             ('C1=CC=CC=C1', [
-                Tok(smiles.ATOM, 'C'),
-                Tok(smiles.DIGIT, 1),
-                Tok(smiles.BOND, '='),
-                Tok(smiles.ATOM, 'C'),
-                Tok(smiles.ATOM, 'C'),
-                Tok(smiles.BOND, '='),
-                Tok(smiles.ATOM, 'C'),
-                Tok(smiles.ATOM, 'C'),
-                Tok(smiles.BOND, '='),
-                Tok(smiles.ATOM, 'C'),
-                Tok(smiles.DIGIT, 1),
+                Tok(tokens.ATOM, 'C'),
+                Tok(tokens.DIGIT, 1),
+                Tok(tokens.BOND, '='),
+                Tok(tokens.ATOM, 'C'),
+                Tok(tokens.ATOM, 'C'),
+                Tok(tokens.BOND, '='),
+                Tok(tokens.ATOM, 'C'),
+                Tok(tokens.ATOM, 'C'),
+                Tok(tokens.BOND, '='),
+                Tok(tokens.ATOM, 'C'),
+                Tok(tokens.DIGIT, 1),
             ]),
             ('N[C@](Br)(O)C', [
-                Tok(smiles.ATOM, 'N'),
-                Tok(smiles.LSPAR, '['),
-                Tok(smiles.ATOM, 'C'),
-                Tok(smiles.AT, '@'),
-                Tok(smiles.RSPAR, ']'),
-                Tok(smiles.LPAR, '('),
-                Tok(smiles.ATOM, 'Br'),
-                Tok(smiles.RPAR, ')'),
-                Tok(smiles.LPAR, '('),
-                Tok(smiles.ATOM, 'O'),
-                Tok(smiles.RPAR, ')'),
-                Tok(smiles.ATOM, 'C'),
+                Tok(tokens.ATOM, 'N'),
+                Tok(tokens.LSPAR, '['),
+                Tok(tokens.ATOM, 'C'),
+                Tok(tokens.AT, '@'),
+                Tok(tokens.RSPAR, ']'),
+                Tok(tokens.LPAR, '('),
+                Tok(tokens.ATOM, 'Br'),
+                Tok(tokens.RPAR, ')'),
+                Tok(tokens.LPAR, '('),
+                Tok(tokens.ATOM, 'O'),
+                Tok(tokens.RPAR, ')'),
+                Tok(tokens.ATOM, 'C'),
             ])
         ]
 
@@ -48,4 +48,4 @@ class LexerTestCase(OSmiPyTestCase):
                 self.assertEqual(token.type, lexed[i].type, msg='{} of {}'.format(i, s))
                 self.assertEqual(token.value, lexed[i].value, msg='{} of {}'.format(i, s))
 
-            self.assertEqual(lexed[-1].type, smiles.EOF)
+            self.assertEqual(lexed[-1].type, tokens.EOF)
