+++ tests/tests_smiles_parser.py
@@ -1,0 +1,125 @@
+from tests import OSmiPyTestCase
+
+from osmipy import lexer, smiles_parser, smiles_ast
+
+
+class ParserTestCase(OSmiPyTestCase):
+
+    def test_parser(self):
+        """Test the parser (on a rather simple string)"""
+
+        smiles_parser_obj = smiles_parser.Parser(lexer.Lexer('C(\\Cl)=C/O'))
+
+        # output of the parser
+        r = smiles_ast.Chain(
+            left=smiles_ast.BranchedAtom(
+                atom=smiles_ast.Atom(symbol='C'),
+                branches=[
+                    smiles_ast.Branch(
+                        chain=smiles_ast.Chain(left=smiles_ast.BranchedAtom(atom=smiles_ast.Atom(symbol='Cl'))),
+                        bond=smiles_ast.Bond('\\'))
+                ]
+            ),
+            right=smiles_ast.Chain(
+                left=smiles_ast.BranchedAtom(atom=smiles_ast.Atom(symbol='C')),
+                right=smiles_ast.Chain(left=smiles_ast.BranchedAtom(atom=smiles_ast.Atom(symbol='O'))),
+                bond=smiles_ast.Bond('/')
+            ),
+            bond=smiles_ast.Bond('='))
+
+        rx = smiles_parser_obj.smiles().chain
+
+        # types
+        self.assertEqual(type(r), type(rx))
+        self.assertEqual(type(r.left), type(rx.left))
+        self.assertEqual(type(r.right), type(rx.right))
+        self.assertEqual(type(r.bond), type(rx.bond))
+        self.assertEqual(type(r.left.branches[0]), type(rx.left.branches[0]))
+        self.assertEqual(type(r.left.branches[0].chain), type(rx.left.branches[0].chain))
+        self.assertEqual(type(r.left.branches[0].bond), type(rx.left.branches[0].bond))
+        self.assertEqual(type(r.left.branches[0].chain.left), type(rx.left.branches[0].chain.left))
+        self.assertEqual(type(r.left.branches[0].chain.left.atom), type(rx.left.branches[0].chain.left.atom))
+        self.assertEqual(type(r.left.atom), type(rx.left.atom))
+        self.assertEqual(type(r.right.left), type(rx.right.left))
+        self.assertEqual(type(r.right.right), type(rx.right.right))
+        self.assertEqual(type(r.right.bond), type(rx.right.bond))
+        self.assertEqual(type(r.right.left.atom), type(rx.right.left.atom))
+        self.assertEqual(type(r.right.right.left.atom), type(rx.right.right.left.atom))
+
+        # symbols and hcounts
+        self.assertEqual(r.left.branches[0].bond.symbol, rx.left.branches[0].bond.symbol)
+        self.assertEqual(r.left.branches[0].chain.left.atom.symbol, rx.left.branches[0].chain.left.atom.symbol)  # -Cl
+        self.assertEqual(r.left.branches[0].chain.left.implicit_hcount, 0)
+        self.assertEqual(r.bond.symbol, rx.bond.symbol)
+        self.assertEqual(r.left.atom.symbol, rx.left.atom.symbol)  # sp² C
+        self.assertEqual(r.left.implicit_hcount, 1)
+        self.assertEqual(r.right.bond.symbol, rx.right.bond.symbol)
+        self.assertEqual(r.right.left.atom.symbol, rx.right.left.atom.symbol)  # sp² C
+        self.assertEqual(r.right.right.left.atom.symbol, rx.right.right.left.atom.symbol)  # -OH
+        self.assertEqual(r.right.right.left.implicit_hcount, 1)
+
+        # parent-child
+        self.assertEqual(r.parent, None)
+        self.assertEqual(r.left.parent, r)
+        self.assertEqual(r.right.parent, r)
+        self.assertEqual(r.bond.parent, r)
+        self.assertEqual(r.left.atom.parent, r.left)
+        self.assertEqual(r.left.branches[0].chain.parent, r.left.branches[0])
+        self.assertEqual(r.left.branches[0].bond.parent, r.left.branches[0])
+
+        # An example with ring bonds
+        r = smiles_ast.Chain(
+            left=smiles_ast.BranchedAtom(
+                atom=smiles_ast.Atom(symbol='C'),
+                ring_bonds=[smiles_ast.RingBond(ring_id=1)]),
+            right=smiles_ast.Chain(
+                left=smiles_ast.BranchedAtom(atom=smiles_ast.Atom(symbol='C')),
+                right=smiles_ast.Chain(
+                    left=smiles_ast.BranchedAtom(
+                        atom=smiles_ast.Atom(symbol='C'),
+                        ring_bonds=[smiles_ast.RingBond(ring_id=1)])
+                )))
+
+        rx = smiles_parser.Parser(lexer.Lexer('C1CC1')).smiles().chain
+
+        # types
+        self.assertEqual(type(r), type(rx))
+        self.assertEqual(type(r.left), type(rx.left))
+        self.assertEqual(type(r.left.atom), type(rx.left.atom))
+        self.assertEqual(type(r.right), type(rx.right))
+        self.assertEqual(type(r.right.left), type(rx.right.left))
+        self.assertEqual(type(r.right.left.atom), type(rx.right.left.atom))
+        self.assertEqual(type(r.right.right.left), type(rx.right.right.left))
+        self.assertEqual(type(r.right.right.left.atom), type(rx.right.right.left.atom))
+
+        # symbols and hcount
+        self.assertEqual(r.left.atom.symbol, rx.left.atom.symbol)
+        self.assertEqual(rx.left.implicit_hcount, 2)  # sp³ C
+        self.assertEqual(r.right.left.atom.symbol, rx.right.left.atom.symbol)
+        self.assertEqual(rx.right.left.implicit_hcount, 2)  # sp³ C
+        self.assertEqual(r.right.right.left.atom.symbol, rx.right.right.left.atom.symbol)
+        self.assertEqual(rx.right.right.left.implicit_hcount, 2)  # sp³ C
+
+        # ring bonds
+        self.assertEqual(type(r.left.ring_bonds[0]), type(rx.left.ring_bonds[0]))
+        self.assertEqual(type(r.right.right.left.ring_bonds[0]), type(rx.right.right.left.ring_bonds[0]))
+
+        self.assertEqual(rx.left.ring_bonds[0].ring_id, rx.right.right.left.ring_bonds[0].ring_id)
+        self.assertEqual(rx.left.ring_bonds[0].target, rx.right.right.left)
+        self.assertEqual(rx.left, rx.right.right.left.ring_bonds[0].target)
+
+    def test_parse_ring_bond(self):
+        """Specific problematic ring bond cases (thus raising errors)"""
+
+        wrong_smiles = [
+            'C11',  # same atom
+            'C12CCCCC12',  # same bond twice
+            'C21CCCCC12',  # same bond twice
+            'C12C2CCC1',  # "2" forms a direct pair
+            'C-1CCCCC=1',  # not the same kind of bond
+            'C12CCCCC1',  # unmatched "2"
+        ]
+
+        for s in wrong_smiles:
+            with self.assertRaises(smiles_parser.ParserException, msg=s):
+                smiles_parser.Parser(lexer.Lexer(s)).smiles()
